Stop preallocating buffers in Marshaling.
It is somewhat error-prone, and the Go compiler has allocation
strategies that work out not that badly.

Since options get stored in a map before actually being marshaled out,
we currently have no way of preallocating one large buffer to write a
Packet out to anyway. If/when the Options map looks something like this:

map[OptionCode][]encoding.BinaryMarshaler

We may be able to outsmart the Go compiler's allocation strategies. For
now, I'm not sure it's worth it without some performance benchmarks.

Signed-off-by: Christopher Koch <c@chrisko.ch>